Do abs operation as separate step
Do the abs as a separate step.  We now return the reconstructed data
with and without the abs operator applied.  While it would be nice to
have a non-negativity constraint, applying abs at the end could create
artifacts.  For example images the signa is 3,2,1,0,-1,-2,-1,0,1,2,3
applying abs would give 3,2,1,0,1,2,1,0,1,2,3.  So it would change the
shape of the curve to give a 'false' pulse in the middle.